Fixed chibios_test_telemetry firmware.
The chibios_test_telemetry example was not initializing the
pprz_transport. This was leading to a crash, unfortunately chibios has
no good way of reporting why a crash happened, backtrace is completely
useless because the ResetHandler is not carrying any information about
the process that caused the crash. (at least not with common tools like
GDB)

While at it the test_serial example is using the rx semaphore now to
wait for data to come in, instead of continously sending out the last
byte that came in.